Replace git apply with patch
Hey, thanks a lot from the AUR package.
Unfortunately I've not been able to install it.
After some manual tinkering with `makepkg` I realized that `git apply` actually just skips both patch files because it searches for them from the root of the git repo instead of the current working directory. The same issue is discussed [here](https://stackoverflow.com/questions/42386491/git-apply-skips-patches). I don't really understand that I had this error and you did not. I have `git version 2.34.0`.
The easiest way to fix it for me was to just not use `git` for that and replace it with `patch`.
I confirmed that it builds and installs normally after that.